# contracts/openapi/v1.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "VANTAGE API",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/aircraft/source": {
      "get": {
        "tags": [
          "Aircraft"
        ],
        "operationId": "Aircraft_Source",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftSourceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aircraft": {
      "get": {
        "tags": [
          "Aircraft"
        ],
        "operationId": "Aircraft_Query",
        "parameters": [
          {
            "name": "longitude",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 12.0
            },
            "x-position": 1
          },
          {
            "name": "latitude",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 58.0
            },
            "x-position": 2
          },
          {
            "name": "radiusNm",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 250
            },
            "x-position": 3
          },
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AircraftRecordDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aircraft/observations/{id}": {
      "get": {
        "tags": [
          "Aircraft"
        ],
        "operationId": "Aircraft_Observation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftRecordDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/earthquakes/source": {
      "get": {
        "tags": [
          "Earthquakes"
        ],
        "operationId": "Earthquakes_Source",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarthquakeSourceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/earthquakes": {
      "get": {
        "tags": [
          "Earthquakes"
        ],
        "operationId": "Earthquakes_Query",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarthquakeSnapshotDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/earthquakes/observations/{id}": {
      "get": {
        "tags": [
          "Earthquakes"
        ],
        "operationId": "Earthquakes_Observation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarthquakeRecordDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/earthquakes/entities/{id}/observations": {
      "get": {
        "tags": [
          "Earthquakes"
        ],
        "operationId": "Earthquakes_Versions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "sourceId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EarthquakeRecordDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
...
      "AircraftSourceDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "documentationUrl": {
            "type": "string"
          },
          "termsUrl": {
            "type": "string"
          },
          "attribution": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pollSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "maximumRadiusNm": {
            "type": "integer",
            "format": "int32"
          },
          "resultLimit": {
            "type": "integer",
            "format": "int32"
          },
          "cacheHours": {
            "type": "integer",
            "format": "int32"
          },
          "minimumRadiusNm": {
            "type": "integer",
            "format": "int32"
          },
          "coverage": {
            "type": "string"
          }
        }
      },
      "AircraftRecordDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/EntityDto"
          },
          "observation": {
            "$ref": "#/components/schemas/AircraftObservationDto"
          },
          "identityRule": {
            "type": "string"
          },
          "identityDescription": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "EntityDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "externalIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalIdDto"
            }
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ExternalIdDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "namespace": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "AircraftObservationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "retrievedAt": {
            "type": "string",
            "format": "date-time"
          },
          "geometry": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PointGeometryDto"
              }
            ]
          },
          "locationRole": {
            "type": "string",
            "nullable": true
          },
          "precision": {
            "$ref": "#/components/schemas/PrecisionDto"
          },
          "evidenceClass": {
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/AircraftPropertiesDto"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceDto"
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PointGeometryDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "coordinates": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          },
          "type": {
            "type": "string"
          }
        }
      },
      "PrecisionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "level": {
            "type": "string"
          }
        }
      },
      "AircraftPropertiesDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string"
          },
          "addressNamespace": {
            "type": "string"
          },
          "callsign": {
            "type": "string",
            "nullable": true
          },
          "registration": {
            "type": "string",
            "nullable": true
          },
          "aircraftType": {
            "type": "string",
            "nullable": true
          },
          "sourceType": {
            "type": "string"
          },
          "speedMetresPerSecond": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "barometricAltitudeMetres": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "ellipsoidAltitudeMetres": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "trackDegrees": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "trueHeadingDegrees": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "onGround": {
            "type": "boolean",
            "nullable": true
          },
          "positionObservedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "containmentRadiusMetres": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "mlatFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProvenanceDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "sourceId": {
            "type": "string"
          },
          "sourceRecordId": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string"
          },
          "attribution": {
            "type": "string"
          },
          "licenseRef": {
            "type": "string"
          },
          "rawRef": {
            "type": "string"
          },
          "derivedFrom": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transformVersion": {
            "type": "string"
          },
          "transformDescription": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "EarthquakeSourceDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "documentationUrl": {
            "type": "string"
          },
          "termsUrl": {
            "type": "string"
          },
          "attribution": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pollSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "resultLimit": {
            "type": "integer",
            "format": "int32"
          },
          "cacheHours": {
            "type": "integer",
            "format": "int32"
          },
          "coverage": {
            "type": "string"
          },
          "scopeLabel": {
            "type": "string"
          },
          "staleAfterSeconds": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "EarthquakeSnapshotDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EarthquakeRecordDto"
            }
          },
          "completeness": {
            "$ref": "#/components/schemas/EarthquakeCompletenessDto"
          },
          "source": {
            "$ref": "#/components/schemas/EarthquakeSourceDto"
          }
        }
      },
      "EarthquakeRecordDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/EntityDto"
          },
          "observation": {
            "$ref": "#/components/schemas/EarthquakeObservationDto"
          },
          "identityRule": {
            "type": "string"
          },
          "identityDescription": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "EarthquakeObservationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "retrievedAt": {
            "type": "string",
            "format": "date-time"
          },
          "geometry": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PointGeometryDto"
              }
            ]
          },
          "locationRole": {
            "type": "string",
            "nullable": true
          },
          "precision": {
            "$ref": "#/components/schemas/PrecisionDto"
          },
          "evidenceClass": {
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/EarthquakePropertiesDto"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceDto"
          },
          "supersedesObservationId": {
            "type": "string",
            "nullable": true
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "EarthquakePropertiesDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "format": "int32"
          },
          "magnitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "magnitudeType": {
            "type": "string",
            "nullable": true
          },
          "depthKilometres": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "depthReference": {
            "type": "string"
          },
          "place": {
            "type": "string",
            "nullable": true
          },
          "sourceUpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reviewStatus": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "type": "string",
            "nullable": true
          },
          "network": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "EarthquakeCompletenessDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "returned": {
            "type": "integer",
            "format": "int32"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "truncated": {
            "type": "boolean"
          },
          "coverage": {
            "type": "string"
          },
          "feedGeneratedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "feedRetrievedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "providerCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rejectedCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}